+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#235](https://github.com/os2display/display-api-service/pull/234)
+  - Fixed Eventdatabasen v2 subscription data order by using occurrences endpoint.
 - [#231](https://github.com/os2display/display-api-service/pull/231)
   - Adds new feed source: Eventdatabasen v2.
 - [#233](https://github.com/os2display/display-api-service/pull/233)